requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.0.2
djangorestframework==3.13.1
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
celery==5.1.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
requests==2.26.0
psycopg2-binary==2.8.6
redis==3.5.3
python-telegram-bot==13.10
django-environ==0.8.1
